requirements.txt
django==4.1
djangorestframework==3.15.0
mysqlclient==2.2.4
django-cors-headers==4.3.1
six
requests